requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.9  # Commented out - requires PostgreSQL dev files on Windows
asyncpg==0.29.0  # Use this for async PostgreSQL operations
# Alternative for development: aiosqlite==0.19.0

# Cache
redis==5.0.1
hiredis==2.3.2

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2

# Code Analysis
javalang==0.13.0
gitpython==3.1.41
pygments==2.17.2
# tree-sitter==0.21.0  # Commented out - Python 3.12 compatibility issue
# Use tree-sitter==0.21.3 for Python 3.12+ or remove if not needed

# AI/ML
spacy==3.7.2
transformers==4.36.2
sentence-transformers==2.3.1
torch==2.1.2
numpy==1.26.3

# HTTP Clients
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
pyjwt==2.8.0

# Task Queue
celery==5.3.6
kombu==5.3.5

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.40.0

# Report Generation
reportlab==4.0.9
openpyxl==3.1.2
jinja2==3.1.3
matplotlib==3.8.2
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
tenacity==8.2.3
python-dateutil==2.8.2

# IBM Cloud & watsonx.ai
ibm-cloud-sdk-core==3.19.0
ibm-watson==8.0.0
ibm-watsonx-ai==0.2.6

# API Documentation
pydantic[email]==2.5.3

# Development Tools (optional, can be in requirements-dev.txt)
# pytest==7.4.4
# pytest-asyncio==0.23.3
# pytest-cov==4.1.0
# black==24.1.1
# flake8==7.0.0
# mypy==1.8.0
# pylint==3.0.3